requirements.txt
requests
PyYAML
beautifulsoup4
plotly
flask